% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddModuleignore.R
\name{AddModuleignore}
\alias{AddModuleignore}
\title{Add ignore file to module}
\usage{
AddModuleignore(dir = ".", contents = NULL)
}
\arguments{
\item{dir}{character. Directory where the module file should be placed.}

\item{contents}{character. Manually specify lines to go into ignore file}
}
\description{
This function writes an ignore file that is parsed by
the temporarily archived \code{PublishModule} and describes files that are not to be packaged
up.
}

