% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSettings.R
\name{AddSettings}
\alias{AddSettings}
\title{Add settings file to project}
\usage{
AddSettings(file = "sws.yml", gitignore = TRUE, fields = NULL)
}
\arguments{
\item{file}{character. Name of settings file to write}

\item{gitignore}{logical. Should this file be added to the .gitignore?}

\item{fields}{list. Manually specify the structure of the settings file}
}
\description{
Creates a sample settings file, by default (leading dashes for
  formatting, whitespace in the actual file):

current: qa \cr
all: \cr
--- share: /media/share \cr
qa: \cr
--- certdir: path/to/qa/certs\cr
--- server: https://hqlqasws1.hq.un.fao.org:8181/sws \cr
--- token: abcdef0123456789 \cr
production: \cr
--- certdir: path/to/prod/certs\cr
--- server: https://hqlprswsas1.hq.un.fao.org:8181/sws \cr
--- token: 9876543210fedcba
}

