% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyKey.R
\name{CopyKey}
\alias{CopyKey}
\title{Copy codes from a DatasetKey to clipboard}
\usage{
CopyKey(key)
}
\arguments{
\item{key}{DatasetKey. Object obtained from the \code{faosws} package. It's
the data type of the \code{swsContext.datasets[[1]]} object}
}
\description{
One of the problems with the Statistical Working System is that of token
reuse. Tokens function as both authentication and data source for debugging,
so it can be very tempting for some developers to use the tokens of others.
This is risky as tokens allow authentication as other users and may also be
revoked by that user at any time. This function makes it easier to duplicate
someone else's session so that token reuse isn't necessary. It sequentially
copies codes to the keyboard so that they can be pasted into the web
interface of the Statistical Working System.
}

