% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoswsModules-package.R
\docType{package}
\name{faoswsModules-package}
\alias{faoswsModules}
\alias{faoswsModules-package}
\title{Tools to create and use Statistical Working System packages.}
\description{
The main functions to get one started with the system are the \code{Add...} functions, namely:
\itemize{
\item \code{\link{AddSettings}} - Add sample settings file for a module
\item \code{\link{AddModule}} - Create dummy module to edit
}
}
\author{
Sebastian Campbell
}

