% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublishModule.R
\name{PublishModule}
\alias{PublishModule}
\title{Publish a module as a zip file}
\usage{
PublishModule(modulepath, ignore = file.path(modulepath, ".moduleignore"))
}
\arguments{
\item{modulepath}{character. Module path relative to the project root}

\item{ignore}{character. Location of ignore file}
}
\description{
When ready to upload your module, this function allows you to ignore files in
a .moduleignore file (as well as any files beginning with "."). It then
checks if the current files correspond to the current commit. If not, it
throws an error. If there is no difference, the files are copied to a
temporary folder and a line appended to the main R file with the current
commit hash. They are then zipped up and copied back to the module folder.
}

