% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublishPackage.R
\name{PublishPackage}
\alias{PublishPackage}
\title{Publish a package to remote repository}
\usage{
PublishPackage(version, message = sprintf("Release version \%s", version),
  packpath = ".", destdir = packpath, date = Sys.Date(),
  document = TRUE, check = TRUE, ignorefile = file.path(packpath,
  ".Rbuildignore"))
}
\arguments{
\item{version}{character. Valid R version. There's no need to prefix it with
a 'v'}

\item{message}{character. Message to go in tag label and}

\item{packpath}{character. Package path relative to the project root}

\item{destdir}{character. Destination directory for package By default,
\code{packpath}}

\item{date}{character or Date. Today's date to go in the Date field of the
DESCRIPTION}

\item{document}{logical, whether documentation should be run}

\item{check}{logical whether RCMD check should be run}

\item{ignorefile}{character. Location of .Rbuildignore file}
}
\description{
Creates a tagged commit suitable for the Statistical Working System. The
documentation is checked, followed by a check that there are no uncommitted
files which could end up the the commit. The commit is then made and tagged.
}

