% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSettings.R
\name{ReadSettings}
\alias{ReadSettings}
\title{Read settings file}
\usage{
ReadSettings(file = "sws.yml", server = NULL)
}
\arguments{
\item{file}{character. Name of the settings file.}

\item{server}{character. This overrides the 'current' field in the settings
file.}
}
\description{
A settings file allows a user of the SWS to make changes to the parameters of
a script without changing the script itself. The format itself is very
flexible yaml file, but has three special sections:
}
\details{
\itemize{
\item \emph{current} - This required field indicates the current
server to which the script is directed
\item \emph{global} - This optional file indicates the location of a global
file from which you may fetch settings.
\item \emph{all} - This optional field specifies settings that should always
be uses. It overrides settings in the global file and is overwritten by
specific server settings.}
}
\seealso{
AddSettings
}

