% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddModule.R
\name{AddModule}
\alias{AddModule}
\title{Make module skeleton}
\usage{
AddModule(name = "new_module", dir = sprintf("modules/\%s", name))
}
\arguments{
\item{name}{character. Name of the new module}

\item{dir}{character. Directory in which to place the new module}
}
\description{
Make everything you need to get started on making modules for the Statisical
Working system including:
}
\details{
\itemize{
\item main.R
\item metadata.xml (future release)
\item script subdirectories (future release)
}
}

