% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSettings.R
\name{AddSettings}
\alias{AddSettings}
\title{Add settings file to project}
\usage{
AddSettings(filename = "sws.yml", gitignore = TRUE, fields = NULL)
}
\arguments{
\item{filename}{character.}

\item{gitignore}{character.}

\item{fields}{list.}
}
\description{
Add settings file to project
}

