% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublishPackage.R
\name{PublishPackage}
\alias{PublishPackage}
\title{Publish a package to remote repository}
\usage{
PublishPackage(version, message = sprintf("Release version \%s", version),
  packpath = ".", destdir = packpath, date = Sys.Date(),
  document = TRUE, check = TRUE, ignorefile = file.path(packpath,
  ".Rbuildignore"))
}
\arguments{
\item{version}{character. Valid R version. There's no need to prefix it with
a 'v'}

\item{packpath}{character. Package path relative to the project root}

\item{destdir}{character. Destination directory for package By default,
\code{packpath}}

\item{ignorefile}{character. Location of .Rbuildignore file}
}
\description{
When ready to upload your new package version, this function allows you to
ignore files in a .moduleignore file (as well as any files beginning with
"."). It then checks if the current files correspond to the current commit.
If not, it throws an error. If there is no difference, the files are copied
to a temporary folder and a line appended to the main R file with the current
commit hash. They are then zipped up and copied back to the module folder.
}

